Removing ReactorWithEffects.MutationWithEffect * This implementation was modeled on what we needed to make effects work on ReactorKit, but emitting effects directly is much cleaner implementation. * Upgraded to maven-publish plugin (facilitated in local deploys and testing)
