+++ rxreactor1-android/build.gradle
@@ -2,7 +2,7 @@ apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
 apply plugin: 'jacoco'
 apply plugin: 'org.jetbrains.dokka'
-apply plugin: 'com.github.dcendents.android-maven'
+apply plugin: 'maven-publish'
 apply from: "$rootDir/gradle/library-artifacts.gradle"
 apply from: "$rootDir/gradle/jacoco-android.gradle"
 
@@ -58,4 +58,46 @@ tasks.withType(Test) {
 
 jacoco {
     toolVersion = versions.jacoco
+}
+
+afterEvaluate {
+    publishing {
+        publications {
+            rxreactor1Android(MavenPublication) {
+                groupId = project.group
+                artifactId = archivesBaseName
+
+                from components.release
+                artifact sourcesJar
+                artifact javadocJar
+
+                pom {
+                    name = "RxReactor (RxJava 1) for Android"
+                    description = "A Kotlin framework for a reactive and unidirectional RxJava 1 application architecture"
+                    url = siteUrl
+
+                    licenses {
+                        license {
+                            name = 'BSD 3-Clause License'
+                            url = 'https://github.com/ggrell/RxReactor/blob/master/LICENSE'
+                        }
+                    }
+
+                    developers {
+                        developer {
+                            id = 'ggrell'
+                            name = 'Gyuri Grell'
+                            email = 'ggrell@pobox.com'
+                        }
+                    }
+
+                    scm {
+                        connection = gitUrl
+                        developerConnection = gitUrl
+                        url = siteUrl
+                    }
+                }
+            }
+        }
+    }
 }
