+++ rxreactor1/src/test/kotlin/com/gyurigrell/rxreactor1/ReactorWithMutationEffectsTests.kt
@@ -7,16 +7,16 @@
 
 package com.gyurigrell.rxreactor1
 
-import com.gyurigrell.rxreactor1.ReactorWithEffectsTests.TestReactor.*
-
+import com.gyurigrell.rxreactor1.ReactorWithMutationEffectsTests.TestReactor.*
 import org.junit.Test
 import rx.Observable
 import rx.observers.TestSubscriber
 
 /**
- * Unit tests for [ReactorWithEffects]
+ * Unit tests for [ReactorWithEffects]. These test use the deprecated [ReactorWithEffects.MutationWithEffect] marker
+ * on [Mutation]s which are now obsolete.
  */
-class ReactorWithEffectsTests {
+class ReactorWithMutationEffectsTests {
     @Test
     fun `SimpleAction updates State simpleAction to true`() {
         // Arrange
@@ -100,6 +100,7 @@ class ReactorWithEffectsTests {
         sealed class Mutation : MutationWithEffect<Effect> {
             object SimpleActionMutation : Mutation()
             data class ActionWithValueMutation(val theValue: String) : Mutation()
+            data class FireEffect(override val effect: Effect) : Mutation()
         }
 
         data class State(
@@ -113,27 +114,22 @@ class ReactorWithEffectsTests {
         }
 
         override fun mutate(action: Action): Observable<Mutation> = when (action) {
-            is Action.SimpleAction ->
-                Observable.just(Mutation.SimpleActionMutation)
+            is Action.SimpleAction -> Observable.just(Mutation.SimpleActionMutation)
 
-            is Action.ActionWithValue ->
-                Observable.just(Mutation.ActionWithValueMutation(action.theValue))
+            is Action.ActionWithValue -> Observable.just(Mutation.ActionWithValueMutation(action.theValue))
 
-            is Action.ActionFiresEffectOne -> {
-                emitEffect(Effect.EffectOne)
-                Observable.empty() // No mutations
-            }
+            is Action.ActionFiresEffectOne -> Observable.just(Mutation.FireEffect(Effect.EffectOne))
 
-            is Action.ActionFiresEffectWithValue -> {
-                emitEffect(Effect.EffectWithValue(action.theValue))
-                Observable.empty() // No mutations
-            }
+            is Action.ActionFiresEffectWithValue ->
+                Observable.just(Mutation.FireEffect(Effect.EffectWithValue(action.theValue)))
         }
 
         override fun reduce(state: State, mutation: Mutation): State = when (mutation) {
             is Mutation.SimpleActionMutation -> state.copy(simpleAction = true)
 
             is Mutation.ActionWithValueMutation -> state.copy(actionWithValue = mutation.theValue)
+
+            is Mutation.FireEffect -> state // This will never happen, but need to be exhaustive
         }
     }
 }
