+++ rxreactor3-android/build.gradle
@@ -1,7 +1,8 @@
 apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
+apply plugin: 'jacoco'
 apply plugin: 'org.jetbrains.dokka'
-apply plugin: 'com.github.dcendents.android-maven'
+apply plugin: 'maven-publish'
 apply from: "$rootDir/gradle/library-artifacts.gradle"
 apply from: "$rootDir/gradle/jacoco-android.gradle"
 
@@ -49,4 +50,54 @@ dependencies {
 
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+}
+
+tasks.withType(Test) {
+    jacoco.includeNoLocationClasses = true
+}
+
+jacoco {
+    toolVersion = versions.jacoco
+}
+
+afterEvaluate {
+    publishing {
+        publications {
+            rxreactor3Android(MavenPublication) {
+                groupId = project.group
+                artifactId = archivesBaseName
+
+                from components.release
+                artifact sourcesJar
+                artifact javadocJar
+
+                pom {
+                    name = "RxReactor (RxJava 3) for Android"
+                    description = "A Kotlin framework for a reactive and unidirectional RxJava 3 application architecture"
+                    url = siteUrl
+
+                    licenses {
+                        license {
+                            name = 'BSD 3-Clause License'
+                            url = 'https://github.com/ggrell/RxReactor/blob/master/LICENSE'
+                        }
+                    }
+
+                    developers {
+                        developer {
+                            id = 'ggrell'
+                            name = 'Gyuri Grell'
+                            email = 'ggrell@pobox.com'
+                        }
+                    }
+
+                    scm {
+                        connection = gitUrl
+                        developerConnection = gitUrl
+                        url = siteUrl
+                    }
+                }
+            }
+        }
+    }
 }
